Real .NET Core platform support
SQLitePCL.pretty had SQLitePCLRaw.core dependency targeted w/o netcore45, that's why it was impossible to install it in .NET Core, even w/ correct .nuspec (netcore45 already presented). This commit fix it, but it'll be a good idea to add support for other profiles supported by SQLitePCLRaw.core too (sl15 and so on).